Fix Lodestone NPE

+++ bot/commands/lodestone/src/main/kotlin/Handler.kt
@@ -43,12 +43,20 @@ class Handler: RequestStreamHandler {
         val event: Interaction<ApplicationCommandData> =
             json.decodeFromString(inputStream.readAllBytes().decodeToString())
 
-        when (event.data!!.options!![0].name) {
-            "card", "Lodestone: Get character card" -> card(event)
-            "link" -> link(event)
-            "unlink" -> unlink(event)
-            "portrait", "Lodestone: Get character portrait" -> portrait(event)
-            "profile", "Lodestone: Get character profile" -> profile(event)
+        if (event.type == 1) {
+            when (event.data!!.options!![0].name) {
+                "card" -> card(event)
+                "link" -> link(event)
+                "unlink" -> unlink(event)
+                "portrait" -> portrait(event)
+                "profile" -> profile(event)
+            }
+        } else if (event.type == 2) {
+            when (event.data!!.name) {
+                "Character card" -> card(event)
+                "Character portrait" -> portrait(event)
+                "Character profile" -> profile(event)
+            }
         }
     }
 }
